+++ PF/arquivos_disponibilizados/Enemy.h
@@ -18,7 +18,7 @@ __BEGIN_API
 class Enemy : public Spaceship {
 
   public:
-    Enemy(ALLEGRO_COLOR color, int size, int speed, int maxLife, Point position, ALLEGRO_COLOR missileColor, ALLEGRO_COLOR laserColor, int missileDelay, int laserDelay);
+    Enemy(ALLEGRO_COLOR color, int size, double speed, int maxLife, Point position, ALLEGRO_COLOR missileColor, ALLEGRO_COLOR laserColor, int missileDelay, int laserDelay);
     ~Enemy() {  };
 
     virtual void draw() {  };
